CVP-1596: Refactor & add more tests
