+++ src/tests/searchParameters/search_by_date/fixtures/Observation/observation2.json
@@ -1,0 +1,82 @@
+{
+  "resourceType": "Observation",
+  "id": "c4171fee-aaa6-4526-a377-ea39860a5c76",
+  "meta": {
+    "source": "https://www.icanbwell.com",
+    "security": [
+      {
+        "system": "https://www.icanbwell.com/owner",
+        "code": "bwell"
+      },
+      {
+        "system": "https://www.icanbwell.com/access",
+        "code": "bwell"
+      }
+    ]
+  },
+  "status": "final",
+  "category": [
+    {
+      "coding": [
+        {
+          "system": "http://terminology.hl7.org/CodeSystem/observation-category",
+          "code": "vital-signs",
+          "display": "Vital Signs"
+        }
+      ]
+    }
+  ],
+  "code": {
+    "coding": [
+      {
+        "system": "http://loinc.org",
+        "code": "96607-7",
+        "display": "Blood pressure panel mean systolic and mean diastolic"
+      }
+    ]
+  },
+  "subject": {
+    "reference": "Patient/1"
+  },
+  "effectivePeriod": {
+    "start": "2020-10-15",
+    "end": "2020-11-18"
+  },
+  "issued": "2019-10-16T12:12:29-10:00",
+  "component": [
+    {
+      "code": {
+        "coding": [
+          {
+            "system": "http://loinc.org",
+            "code": "96608-5",
+            "display": "Systolic blood pressure mean"
+          }
+        ]
+      },
+      "valueQuantity": {
+        "value": 120,
+        "unit": "mm[Hg]",
+        "system": "http://unitsofmeasure.org",
+        "code": "mm[Hg]"
+      }
+    },
+    {
+      "code": {
+        "coding": [
+          {
+            "system": "http://loinc.org",
+            "code": "96609-3",
+            "display": "Diastolic blood pressure mean"
+          }
+        ]
+      },
+      "valueQuantity": {
+        "value": 80,
+        "unit": "mm[Hg]",
+        "system": "http://unitsofmeasure.org",
+        "code": "mm[Hg]"
+      }
+    }
+  ]
+}
